refactor: move getPosInfo function
